fix 64 bit packing

+++ Event/EventPacker/Event/PackedMCRichDigitSummary.h
@@ -4,7 +4,6 @@
 
 // STL
 #include <string>
-#include <cstdint>
 
 // Kernel
 #include "Event/StandardPacker.h"
@@ -30,8 +29,8 @@ namespace LHCb
    */
   struct PackedMCRichDigitSummary
   {
-    int history{0};
-    uint64_t richSmartID{0};
+    int       history{0};
+    long long richSmartID{0};
     long long mcParticle{-1};
   };
