+++ Event/EventPacker/Event/PackedMCRichHit.h
@@ -4,7 +4,6 @@
 
 // STL
 #include <string>
-#include <cstdint>
 
 // Kernel
 #include "Event/StandardPacker.h"
@@ -30,11 +29,11 @@ namespace LHCb
    */
   struct PackedMCRichHit
   {
-    int   x{0},y{0},z{0};
-    int   energy{0};
-    int   tof{0};
-    uint64_t sensDetID{0};
-    int   history{0};
+    int       x{0},y{0},z{0};
+    int       energy{0};
+    int       tof{0};
+    long long sensDetID{0};
+    int       history{0};
     long long mcParticle{-1};
   };
